requirements.txt
flask==3.1.0
openai-whisper==20240930
sounddevice==0.5.1
numpy==2.0.2
pyttsx3==2.98
torch==2.5.1
tqdm==4.67.1
